refactor: common/moduleconfig -> internal/moduleconfig
