+++ internal/buildengine/module.go
@@ -1,7 +1,7 @@
 package buildengine
 
 import (
-	"github.com/TBD54566975/ftl/common/moduleconfig"
+	"github.com/TBD54566975/ftl/internal/moduleconfig"
 	"github.com/TBD54566975/ftl/internal/reflect"
 )
