+++ internal/buildengine/watch_test.go
@@ -11,9 +11,9 @@ import (
 	"github.com/alecthomas/assert/v2"
 	"github.com/alecthomas/types/pubsub"
 
-	"github.com/TBD54566975/ftl/common/moduleconfig"
 	. "github.com/TBD54566975/ftl/internal/buildengine"
 	"github.com/TBD54566975/ftl/internal/log"
+	"github.com/TBD54566975/ftl/internal/moduleconfig"
 )
 
 const pollFrequency = time.Millisecond * 500
