+++ internal/moduleconfig/moduleconfig_test.go
