+++ internal/moduleconfig/testdata/imports.go
